Move exports components based on feedback
